requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
kafka-python==2.2.15
pytz==2025.2
requests==2.32.4
urllib3==1.26.20
